♻️ Restructure the SVGAII driver as a GraphicDevice bound to its PciDevice
SvgaIIDriver now implements the GraphicDevice contract (like GopDriver)
and receives the PciDevice it binds to instead of self-locating through
PciManager in its constructor. VMWareSVGAII3D shrinks to the SVGA3D
command layer built on top of the driver, created only when the device
negotiates 3D support. SVGAII3DCanvas draws through the contract instead
of poking VideoMemory with its own byte math, and FullScreenCanvas picks
the canvas from a single PCI lookup instead of the HasSvgaII() probe.
